refactor(auth): extract paths/extraction/headers into _auth subpackage (tier-10 PR-B-low)
Continues the tier-10 auth decomposition by lifting three orthogonal concerns
out of src/notebooklm/auth.py into dedicated private modules:

- _auth/paths.py: env-var name constants (NOTEBOOKLM_REFRESH_CMD_ENV,
  NOTEBOOKLM_REFRESH_CMD_USE_SHELL_ENV, _REFRESH_ATTEMPTED_ENV,
  NOTEBOOKLM_DISABLE_KEEPALIVE_POKE_ENV) + _rotation_lock_path helper.
- _auth/extraction.py: WIZ field token extraction (extract_wiz_field,
  extract_csrf_from_html, extract_session_id_from_html, _build_wiz_field_patterns,
  _safe_url).
- _auth/headers.py: _resolve_token_route_kwargs (intentionally small —
  the bulk of authuser/header helpers already live in _auth/account.py).

auth.py now re-exports each moved symbol via direct module-level assignment
(name = _auth_X.name) — same pattern already used for _auth/account.py and
_auth/cookies.py, so the public surface (notebooklm.auth.X) and the
white-box surface (notebooklm.auth._X) keep resolving identically.

Constraints honored:
- AuthTokens dataclass body untouched (tier-9 E's field(repr=False) +
  custom __repr__ redaction stays exactly as merged in #b42c900).
- extract_email_from_html NOT moved — it already lives in _auth/account.py:79
  routed via _AUTH_ACCOUNT_FACADE_NAMES.
- _AuthFacadeModule untouched (PR-B-high territory).
- _auth/keepalive / _auth/refresh internals untouched (PR-B-high territory).
