+++ src/notebooklm/_auth/__init__.py
@@ -1,1 +1,5 @@
 """Private authentication implementation package."""
+
+from . import extraction, headers, paths
+
+__all__ = ["extraction", "headers", "paths"]
