+++ src/notebooklm/auth.py
@@ -31,7 +31,6 @@ import asyncio
 import contextlib
 import logging
 import os
-import re
 import shlex
 import subprocess
 import sys
@@ -44,17 +43,18 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from types import ModuleType
 from typing import Any, TypeAlias
-from urllib.parse import urlparse
 
 import httpx
 
 from ._auth import account as _auth_account
 from ._auth import cookie_policy as _cookie_policy
 from ._auth import cookies as _auth_cookies
+from ._auth import extraction as _auth_extraction
+from ._auth import headers as _auth_headers
+from ._auth import paths as _auth_paths
 from ._auth import storage as _auth_storage
 from ._env import get_base_url
-from ._url_utils import contains_google_auth_redirect, is_google_auth_redirect
-from .exceptions import AuthExtractionError
+from ._url_utils import is_google_auth_redirect
 from .paths import get_storage_path, resolve_profile
 
 logger = logging.getLogger(__name__)
@@ -466,182 +466,16 @@ class AuthTokens:
         )
 
 
-def _build_wiz_field_patterns(key: str) -> list[re.Pattern[str]]:
-    """Build the ordered list of regex patterns used to locate a Wiz field.
-
-    Patterns are tried in priority order: canonical double-quoted form first,
-    then single-quoted, then HTML-escaped. Each pattern captures the value
-    (which may be empty — empty tokens are legitimate, not a drift signal).
-
-    All three variants tolerate backslash-escaped delimiters inside the value
-    so JSON-style escapes like ``"key":"a\\"b"`` parse correctly. The inner
-    character class ``[^"\\\\]*(?:\\\\.[^"\\\\]*)*`` is the standard
-    "string with escapes" idiom: consume runs of non-quote/non-backslash
-    chars, optionally followed by an escape pair (``\\.``) and another run.
-
-    The HTML-escaped variant uses a tempered-dot lookahead so the capture
-    stops only at a literal ``&quot;`` terminator (not at any ``&`` — values
-    legitimately contain ``&amp;`` and similar entities).
-
-    Whitespace tolerance (``\\s*``) around the colon mirrors the original
-    ``extract_csrf_from_html`` regex so we don't regress.
-    """
-    escaped = re.escape(key)
-    return [
-        # 1. Canonical double-quoted: "key":"value"  (or  "key" : "value")
-        #    Captures escaped quotes: "key":"a\"b" -> a\"b
-        re.compile(rf'"{escaped}"\s*:\s*"([^"\\]*(?:\\.[^"\\]*)*)"'),
-        # 2. Single-quoted variant: 'key':'value' with escaped-quote support.
-        re.compile(rf"'{escaped}'\s*:\s*'([^'\\]*(?:\\.[^'\\]*)*)'"),
-        # 3. HTML-escaped: &quot;key&quot;:&quot;value&quot;
-        #    Tempered dot so the value can contain other entities like &amp;.
-        re.compile(rf"&quot;{escaped}&quot;\s*:\s*&quot;((?:(?!&quot;).)*)&quot;"),
-    ]
-
-
-def extract_wiz_field(html: str, key: str, *, strict: bool = True) -> str | None:
-    """Extract a ``WIZ_global_data[key]`` value from a NotebookLM HTML response.
-
-    NotebookLM (and other Google products) embed a JavaScript object literal
-    named ``WIZ_global_data`` in the page chrome. Tokens like ``SNlM0e``
-    (CSRF) and ``FdrFJe`` (session ID) live inside that object. This helper
-    is the single place that knows how to parse the embedding so all callers
-    benefit from the same drift tolerance and diagnostics.
-
-    Tolerated input variants, tried in priority order:
-
-    1. Canonical double-quoted ``"key":"value"`` (typical raw HTML).
-    2. Single-quoted ``'key':'value'`` (rare, observed in some debug renders).
-    3. HTML-escaped ``&quot;key&quot;:&quot;value&quot;`` (when the script
-       block is rendered inside an attribute or escaped fragment).
-
-    Empty values are passed through verbatim: ``"SNlM0e":""`` returns the
-    empty string. Some Google endpoints legitimately emit empty tokens (e.g.
-    for unauthenticated probes) and the caller — not this helper — should
-    decide whether an empty value is acceptable.
-
-    Args:
-        html: The page HTML to search.
-        key: Field name to extract from ``WIZ_global_data``.
-        strict: When True (default) and no pattern matches, raise
-            :class:`AuthExtractionError` with a sanitized preview. When False,
-            return ``None`` on drift so callers can fall back gracefully.
-
-    Returns:
-        The extracted value (possibly empty), or ``None`` when ``strict=False``
-        and no pattern matched.
-
-    Raises:
-        AuthExtractionError: ``strict=True`` and the key was not found.
-    """
-    for pattern in _build_wiz_field_patterns(key):
-        match = pattern.search(html)
-        if match is not None:
-            return match.group(1)
-    if strict:
-        raise AuthExtractionError(key, html)
-    return None
-
-
-def _safe_url(url: str) -> str:
-    """Return ``url`` stripped of credential-shaped parts for error display.
-
-    Auth-handshake URLs can carry credentials in three positions, all of
-    which we strip:
-
-    * **Query string** — ``f.sid=...``, ``continue=...``, ``access_token=...``.
-    * **Fragment** — OAuth implicit-flow tokens (``#access_token=...``).
-    * **Userinfo** — ``https://TOKEN@host/...`` shapes; ``parsed.netloc``
-      preserves the userinfo, so we rebuild from ``hostname`` + optional
-      port instead of trusting ``netloc`` directly.
-
-    The surviving ``scheme://host[:port]/path`` is enough context for an
-    operator to recognize which endpoint failed without leaking session
-    state. Empty input passes through verbatim so error messages with the
-    default ``final_url=""`` still render cleanly instead of degenerating to
-    ``"://"``.
-    """
-    if not url:
-        return ""
-    parsed = urlparse(url)
-    # hostname strips userinfo; port survives separately. Both can be None
-    # on malformed input, in which case we degrade gracefully to "scheme:///path".
-    host = parsed.hostname or ""
-    netloc = f"{host}:{parsed.port}" if parsed.port is not None else host
-    return f"{parsed.scheme}://{netloc}{parsed.path}"
-
-
-def extract_csrf_from_html(html: str, final_url: str = "") -> str:
-    """
-    Extract CSRF token (SNlM0e) from NotebookLM page HTML.
-
-    The CSRF token is embedded in the page's WIZ_global_data JavaScript object.
-    It's required for all RPC calls to prevent cross-site request forgery.
-
-    Args:
-        html: Page HTML content from notebooklm.google.com
-        final_url: The final URL after redirects (for error messages)
-
-    Returns:
-        CSRF token value (typically starts with "AF1_QpN-")
-
-    Raises:
-        ValueError: Preserved for backward compatibility — raised both when
-            redirected to a Google login page and when the token is missing
-            from a non-redirect response. Existing callers and tests rely on
-            the ``"CSRF token not found"`` / ``"Authentication expired"``
-            message substrings, so we intentionally keep the legacy type.
-            Internally we delegate to :func:`extract_wiz_field` so the regex
-            matrix (double-quoted / single-quoted / HTML-escaped) is shared.
-    """
-    # Tolerant extraction via the unified helper — accepts canonical,
-    # single-quoted, and HTML-escaped variants of the WIZ_global_data field.
-    token = extract_wiz_field(html, "SNlM0e", strict=False)
-    if token is not None:
-        return token
-    # Drift path: differentiate "auth expired" from "shape changed" because
-    # the remediation differs (re-login vs file a bug).
-    if is_google_auth_redirect(final_url) or contains_google_auth_redirect(html):
-        raise ValueError(
-            "Authentication expired or invalid. Run 'notebooklm login' to re-authenticate."
-        )
-    raise ValueError(
-        f"CSRF token not found in HTML. Final URL: {_safe_url(final_url)}\n"
-        "This may indicate the page structure has changed."
-    )
-
-
-def extract_session_id_from_html(html: str, final_url: str = "") -> str:
-    """
-    Extract session ID (FdrFJe) from NotebookLM page HTML.
-
-    The session ID is embedded in the page's WIZ_global_data JavaScript object.
-    It's passed in URL query parameters for RPC calls.
-
-    Args:
-        html: Page HTML content from notebooklm.google.com
-        final_url: The final URL after redirects (for error messages)
-
-    Returns:
-        Session ID value
-
-    Raises:
-        ValueError: Preserved for backward compatibility — raised both when
-            redirected to a Google login page and when the session ID is
-            missing from a non-redirect response. See
-            :func:`extract_csrf_from_html` for the rationale.
-    """
-    sid = extract_wiz_field(html, "FdrFJe", strict=False)
-    if sid is not None:
-        return sid
-    if is_google_auth_redirect(final_url) or contains_google_auth_redirect(html):
-        raise ValueError(
-            "Authentication expired or invalid. Run 'notebooklm login' to re-authenticate."
-        )
-    raise ValueError(
-        f"Session ID not found in HTML. Final URL: {_safe_url(final_url)}\n"
-        "This may indicate the page structure has changed."
-    )
+# WIZ field token extraction (CSRF, session ID, generic WIZ data) lives in
+# ``notebooklm._auth.extraction``. Re-exported here so the public surface
+# (``notebooklm.auth.extract_csrf_from_html`` etc., listed in ``__all__``) and
+# white-box test affordances (``_safe_url``, ``_build_wiz_field_patterns``)
+# keep resolving against ``notebooklm.auth``.
+_build_wiz_field_patterns = _auth_extraction._build_wiz_field_patterns
+_safe_url = _auth_extraction._safe_url
+extract_csrf_from_html = _auth_extraction.extract_csrf_from_html
+extract_session_id_from_html = _auth_extraction.extract_session_id_from_html
+extract_wiz_field = _auth_extraction.extract_wiz_field
 
 
 Account = _auth_account.Account
@@ -706,9 +540,14 @@ def load_auth_from_storage(path: Path | None = None) -> dict[str, str]:
     return extract_cookies_from_storage(storage_state)
 
 
-NOTEBOOKLM_REFRESH_CMD_ENV = "NOTEBOOKLM_REFRESH_CMD"
-NOTEBOOKLM_REFRESH_CMD_USE_SHELL_ENV = "NOTEBOOKLM_REFRESH_CMD_USE_SHELL"
-_REFRESH_ATTEMPTED_ENV = "_NOTEBOOKLM_REFRESH_ATTEMPTED"
+# Env-var name constants live in ``notebooklm._auth.paths``. Re-exported so
+# both the public surface (``NOTEBOOKLM_REFRESH_CMD_ENV``,
+# ``NOTEBOOKLM_REFRESH_CMD_USE_SHELL_ENV`` — listed in ``__all__``) and the
+# white-box surface (``_REFRESH_ATTEMPTED_ENV``, used by tests) keep resolving
+# against ``notebooklm.auth``.
+NOTEBOOKLM_REFRESH_CMD_ENV = _auth_paths.NOTEBOOKLM_REFRESH_CMD_ENV
+NOTEBOOKLM_REFRESH_CMD_USE_SHELL_ENV = _auth_paths.NOTEBOOKLM_REFRESH_CMD_USE_SHELL_ENV
+_REFRESH_ATTEMPTED_ENV = _auth_paths._REFRESH_ATTEMPTED_ENV
 # The ContextVar prevents same-task retry loops in the parent process. The env
 # flag is passed only to child refresh commands so recursive CLI calls skip refresh.
 _REFRESH_ATTEMPTED_CONTEXT: ContextVar[bool] = ContextVar(
@@ -1250,7 +1089,10 @@ _KEEPALIVE_ROTATE_HEADERS = {
 # the in-house experiments referenced in #345; kept in one place so it can be
 # changed if Google ever changes the contract.
 _KEEPALIVE_ROTATE_BODY = '[000,"-0000000000000000000"]'
-NOTEBOOKLM_DISABLE_KEEPALIVE_POKE_ENV = "NOTEBOOKLM_DISABLE_KEEPALIVE_POKE"
+# Env-var name lives in ``notebooklm._auth.paths``; re-exported to keep
+# ``notebooklm.auth.NOTEBOOKLM_DISABLE_KEEPALIVE_POKE_ENV`` (a public-surface
+# name listed in ``__all__``) resolving against this module.
+NOTEBOOKLM_DISABLE_KEEPALIVE_POKE_ENV = _auth_paths.NOTEBOOKLM_DISABLE_KEEPALIVE_POKE_ENV
 _KEEPALIVE_POKE_TIMEOUT = 15.0
 # Skip the poke if storage_state.json was rewritten within this window — protects
 # accounts.google.com from rapid CLI loops (e.g. 10 sequential `notebooklm`
@@ -1343,15 +1185,10 @@ def _try_claim_rotation(storage_path: Path | None) -> bool:
         return True
 
 
-def _rotation_lock_path(storage_path: Path | None) -> Path | None:
-    """Sibling sentinel used by ``_poke_session`` for cross-process coordination.
-
-    Distinct from the ``.storage_state.json.lock`` used by ``save_cookies_to_storage``
-    so a long-running save doesn't block rotations or vice versa.
-    """
-    if storage_path is None:
-        return None
-    return storage_path.with_name(f".{storage_path.name}.rotate.lock")
+# Rotation sentinel path lives in ``notebooklm._auth.paths``; re-exported so
+# tests and internal callers keep resolving the white-box affordance against
+# ``notebooklm.auth``.
+_rotation_lock_path = _auth_paths._rotation_lock_path
 
 
 @contextlib.contextmanager
@@ -1618,28 +1455,10 @@ async def _fetch_tokens_with_jar(
         return csrf, session_id
 
 
-def _resolve_token_route_kwargs(
-    storage_path: Path | None,
-    *,
-    authuser: int | None,
-    account_email: str | None,
-) -> dict[str, Any]:
-    """Resolve token-fetch routing while preserving explicit caller intent."""
-    explicit_authuser = authuser is not None
-    resolved_authuser = get_authuser_for_storage(storage_path) if authuser is None else authuser
-    if account_email is not None:
-        resolved_account_email = account_email
-    elif explicit_authuser:
-        resolved_account_email = None
-    else:
-        resolved_account_email = get_account_email_for_storage(storage_path)
-
-    route_kwargs: dict[str, Any] = {"authuser": resolved_authuser}
-    if resolved_account_email is not None:
-        route_kwargs["account_email"] = resolved_account_email
-    if explicit_authuser:
-        route_kwargs["force_authuser_query"] = True
-    return route_kwargs
+# Token-route resolver lives in ``notebooklm._auth.headers``; re-exported so
+# internal callers (``fetch_tokens``, ``fetch_tokens_with_domains``) and any
+# white-box tests keep resolving the helper against ``notebooklm.auth``.
+_resolve_token_route_kwargs = _auth_headers._resolve_token_route_kwargs
 
 
 async def fetch_tokens(
